Added JWT as a service
