+++ app/services/authentication_token.rb
@@ -1,0 +1,49 @@
+module AuthenticationToken
+
+  def encode_token(payload)
+    JWT.encode(payload, secret, algorithm)
+  end
+
+  def auth_header
+    # { Authorization: 'Bearer <token>' }
+    request.headers['Authorization']
+  end
+
+  def decoded_token
+    return unless auth_header
+
+    token = auth_header.split[1]
+    # header: { 'Authorization': 'Bearer <token>' }
+    begin
+      JWT.decode(token, secret, true, algorithm: algorithm)
+    rescue JWT::DecodeError
+      nil
+    end
+  end
+
+  def logged_in_user
+    return unless decoded_token
+
+    user_id = decoded_token[0]['user_id']
+    @user = User.find_by(id: user_id)
+  end
+
+  def logged_in?
+    !!logged_in_user
+  end
+
+  def authorized
+    render json: { message: 'Please log in' }, status: :unauthorized unless logged_in?
+  end
+
+  private
+
+  def secret
+    # Rails.application.secrets.secret_key_base
+    '123456'
+  end
+
+  def algorithm
+    'HS256'
+  end
+end
